requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
pydantic
pydantic-settings
torch
torchvision
pillow
python-multipart
aiofiles
numpy
pydicom
dependency-injector

# Authentication & Authorization
casbin
casbin-sqlalchemy-adapter
python-jose[cryptography]
argon2-cffi-bindings
pyotp
qrcode[pil]
authlib
slowapi
httpx
pwdlib[argon2]